test: update Pluto tests to use new Pluto
